requirements.txt
absl-py==0.9.0
astor==0.8.1
bleach==1.5.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
enum34==1.1.6
Flask==1.1.1
gast==0.2.2
google-pasta==0.1.8
grpcio==1.26.0
gunicorn==20.0.4
h5py==2.10.0
html5lib==0.9999999
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
numpy==1.18.1
olefile==0.46
opt-einsum==3.1.0
Pillow==4.3.0
protobuf==3.11.2
pytz==2019.3
PyYAML==5.3
requests==2.22.0
scipy==1.4.1
six==1.14.0
tensorboard==1.15.0
tensorflow==1.5.0
tensorflow-estimator==1.15.1
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
urllib3==1.25.8
Werkzeug==0.16.0
wrapt==1.11.2
